% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getChemicalClass.R
\name{getChemicalClass}
\alias{getChemicalClass}
\title{getChemicalClass}
\usage{
getChemicalClass(chemical.names)
}
\arguments{
\item{chemical.names}{vector of molecule names}
}
\description{
function for getting the chemical class of a molecule from its name
}
\examples{
getChemicalClass(c("ethanol", "acetic acid", "isopentyl acetate"))
}

