% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getIdentifierDoOR.R
\name{getIdentifierDoOR}
\alias{getIdentifierDoOR}
\title{getIdentifierDoOR}
\usage{
getIdentifierDoOR(query, representation = "stdinchikey", first = T,
  verbose = T, ...)
}
\arguments{
\item{query}{}

\item{representation}{}

\item{first}{}

\item{verbose}{}

\item{...}{}
}
\description{
lookup chemical identifiers using the webchem package
}

