% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/showDuplicates.R
\name{showDuplicates}
\alias{showDuplicates}
\title{showDuplicates}
\usage{
showDuplicates(data, check.column = "InChIKey",
  gsub = c("\\\\s|-|\\\\+|,|\\\\(|\\\\)|\\\\*|#"))
}
\arguments{
\item{data}{}

\item{check.column}{}

\item{gsub}{}
}
\description{
look for duplicated odor identifiers and show them
}

